refactor: move reservations screen to dedicated folder and add confirmation modal
